attempt to decompile sub_GAME_7F01CC94

+++ src/game/blood_decrypt.h
@@ -7,6 +7,6 @@ u8 *decrypt_bleeding_animation_data(u8*, s32, s32, u8*, s8*);
 void sub_GAME_7F01D16C(u8*, s32, s32, u8*);
 void sub_GAME_7F01D02C(u8*, s32, u8*);
 void sub_GAME_7F01CEEC(u8*, s32, u8*);
-void sub_GAME_7F01CC94(u8*, s32, u8*);
+void sub_GAME_7F01CC94(u8*, u16, u8*);
 
 #endif
